+++ Gdax/MerketData/IMarketDataService.cs
@@ -1,0 +1,9 @@
+﻿using Gdax.MarketData.Products;
+
+namespace Gdax.MerketData
+{
+	public interface IMarketDataService
+	{
+		IProductService Products { get; }
+	}
+}